Adds ESP32 support and updates README to include SetHardware details
The Heltec WiFi LoRa 32 module has been added, which required a few
modifications to the EEPROM save/restore calls.  The ESP32 library is
very close to the Arduino EEPROM library, but still needed a couple of
its own tweaks.

The README.md has been updated to include details on using the
SetHardware functions, as well as pointers to the KISS protocol and two
more APRS/Packet applications that have been useful.